fix: real to decimal

+++ win64-cmd/dump-data/2-wakatime-psql.sql
@@ -9,7 +9,7 @@ CREATE TABLE stash.wakatime_heartbeats (
   "branch" VARCHAR NULL,
   "category" VARCHAR NULL,
   "type" VARCHAR NULL,
-  "time" REAL NULL,
+  "time" DECIMAL NULL,
   "dependencies" VARCHAR NULL,
   "entity" VARCHAR NULL,
   "language" VARCHAR NULL,
